+++ NEWS.md
@@ -9,6 +9,12 @@
 * Fix Rails/Test::Unit integration to ensure that the test case classes we are
   re-opening actually exist.
 
+* The `route` matcher now accepts specifying a controller/action pair as a
+  string instead of only a hash (e.g. `route(...).to('posts#index')` instead of
+  `route(...).to(controller: 'posts', action: 'index')`).
+
+* The `ensure_inclusion_of` matcher now works with a decimal column.
+
 # v 2.4.0
 
 * Fix a bug with the `validate_numericality_of` matcher that would not allow the